+++ test/ssl_test.js
@@ -1,117 +1,13 @@
-var tape_test = require('tape');
-var instrumental;
-var originalConfig = require("../exampleConfig.js").config;
-var https = require("https");
-var EventEmitter = require('events').EventEmitter;
+var TestHelper = require("./test_helper");
 var timekeeper = require('timekeeper');
 var path = require("path");
 var util = require("util");
 var fs   = require("fs");
 
-var timedOut = false;
-var timer, config, log;
-
-var test = function(name, testFunction){
-  tape_test(name, function(t){
-    // Clear any state in the backend
-    delete require.cache[require.resolve('../lib/instrumental.js')]
-    instrumental = require("../lib/instrumental.js");
-
-    // Start with a fresh config based off the example with minor modifications
-    config = JSON.parse(JSON.stringify(originalConfig));
-    config.instrumental.key = process.env.INSTRUMENTAL_TEST_TOKEN;
-    config.instrumental.recordCounterRates = false;
-    config.instrumental.host = "collector.instrumentalapp.com";
-    config.debug = true; // allow log verification
-    config.instrumental.metricPrefix = "";
-    config.instrumental.log = function(){
-      // console.warn(util.format.apply(null, arguments));
-      log.push(util.format.apply(null, arguments));
-    };
-
-    // Setup state to check for timeouts polling the API
-    timedOut = false;
-    if (timer) clearTimeout(timer);
-
-    // Reset timekeeper so time behaves normally by default
-    timekeeper.reset();
-
-    // Collect log messages for checking in tests
-    log = [];
-
-    // Run the test
-    testFunction(t);
-  });
-};
-
-var instrumentalLatency = 20000;
-
-function setup(t){
-  t.timeoutAfter(instrumentalLatency*2.2); // needs to be more than Instrumental latency, a little more than the timeout below
-
-  // Setup something to stop the checkForMetric loop so node exits.
-  timer = setTimeout(function(){
-    timedOut = true;
-  }, instrumentalLatency*2);
-  tape_test.onFinish(function(){
-    clearTimeout(timer);
-  });
-}
-
-function sendMetric(metricName, time, options){
-  if(typeof(options) === 'undefined') options = {};
-  if (!options.skipInit) {
-    dummy_events = new EventEmitter();
-    instrumental.init(now, config, dummy_events)
-  };
-  metrics = {
-    counters: {},
-    counter_rates: {},
-    timer_data: {},
-    gauges: {},
-    sets: {},
-  };
-  metrics.counters[metricName] = 1;
-  dummy_events.emit("flush", time, metrics);
-}
-
-function checkForMetric(metricName, options) {
-  var httpOptions = {
-    hostname: 'instrumentalapp.com',
-    path: '/api/2/metrics/'+metricName,
-    headers: {
-      'X-Instrumental-Token': process.env.INSTRUMENTAL_TEST_TOKEN,
-    }
-  };
-  var req = https.get(httpOptions, function(res){
-    // console.warn('statusCode:', res.statusCode);
-    // console.warn('headers:', res.headers);
-    var body = '';
-    res.on('data', function(chunk) {
-      body += chunk;
-    });
-    res.on('end', function() {
-      // console.warn(body);
-      var data = JSON.parse(body).response.metrics[0].values.data;
-      var last_point = data[data.length-1];
-      var expectedSum = options.expectedSum || 1;
-      if (last_point.s == expectedSum) {
-        options.found();
-      } else {
-        if (timedOut) {
-          options.timeout();
-        } else {
-          setTimeout(function(){checkForMetric(metricName, options)}, 1000);
-        }
-      }
-    });
-  });
-  req.on('error', function(e){
-    console.error(e.message);
-    options.error();
-  });
-}
-
+test = TestHelper.test;
+setup = TestHelper.setup;
+sendMetric = TestHelper.sendMetric;
+checkForMetric = TestHelper.checkForMetric;
 
 test('specifying a valid and working cert bundle works', function(t) {
   setup(t);
@@ -126,8 +22,8 @@ test('specifying a valid and working cert bundle works', function(t) {
     }
   });
 
-  config.instrumental.host = "smoke-collector.instrumentalapp.com";
-  config.instrumental.caCertFiles = certs;
+  TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
+  TestHelper.config.instrumental.caCertFiles = certs;
   now = Math.round(new Date().getTime() / 1000);
   var metricName = "test.metric"+Math.random();
   sendMetric(metricName, oldTime);
@@ -161,15 +57,15 @@ test('specifying a valid but not working cert bundle retries', function(t) {
     }
   });
 
-  config.instrumental.host = "smoke-collector.instrumentalapp.com";
-  config.instrumental.caCertFiles = certs;
+  TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
+  TestHelper.config.instrumental.caCertFiles = certs;
   now = Math.round(new Date().getTime() / 1000);
   var metricName = "test.metric"+Math.random();
   sendMetric(metricName, oldTime);
   sendMetric(metricName, oldTime, {skipInit: true});
   var checkConnectionErrors = function() {
     var connection_errors =
-      log.filter(function(entry){return entry.match(/Client error:/)});
+      TestHelper.log.filter(function(entry){return entry.match(/Client error:/)});
     t.equal(connection_errors.length, 2, "expected 2 connection attemps, 1 retry");
     t.end();
   };
@@ -181,8 +77,8 @@ test('specifying an invalid cert bundle errors', function(t) {
 
   oldTime = Math.round(new Date().getTime() / 1000);
 
-  config.instrumental.host = "smoke-collector.instrumentalapp.com";
-  config.instrumental.caCertFiles = ["non_existent_file"];
+  TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
+  TestHelper.config.instrumental.caCertFiles = ["non_existent_file"];
   var metricName = "test.metric"+Math.random();
   t.throws(function(){
     sendMetric(metricName, oldTime);
@@ -195,14 +91,14 @@ test('node default certs are used by default', function(t) {
 
   oldTime = Math.round(new Date().getTime() / 1000);
 
-  config.instrumental.host = "smoke-collector.instrumentalapp.com";
+  TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
   var metricName = "test.metric"+Math.random();
   sendMetric(metricName, oldTime);
 
   checkForMetric(metricName, {
     found: function(){
       var cert_log_messages =
-        log.filter(function(entry){return entry.match(/Using certs/i)});
+        TestHelper.log.filter(function(entry){return entry.match(/Using certs/i)});
       t.deepEqual(cert_log_messages, ["Using certs: node default"], "expected to use node default certs");
       t.pass();
       t.end();
@@ -228,14 +124,14 @@ test('before cert expiration, fallback to newest cert bundle and stick if node d
 
   var metricName = "test.metric"+Math.random();
 
-  config.instrumental.tlsVariationTimeout = 500;
+  TestHelper.config.instrumental.tlsVariationTimeout = 500;
 
   var timeBetweenSends = 1000;
   var actions = [];
 
   // node default certs
   actions.push(function(){
-    config.instrumental.host = "smoke-collector.instrumentalapp.com";
+    TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
     sendMetric(metricName, realCurrentTime); // using node default certs
   });
   actions.push(function(){
@@ -244,11 +140,11 @@ test('before cert expiration, fallback to newest cert bundle and stick if node d
 
   // New cert bundle
   actions.push(function(){
-    config.instrumental.port = 8000;
+    TestHelper.config.instrumental.port = 8000;
     sendMetric(metricName, realCurrentTime); // node default certs fail
   });
   actions.push(function(){
-    config.instrumental.port = 8001;
+    TestHelper.config.instrumental.port = 8001;
     sendMetric(metricName, realCurrentTime); // bundle
   });
   actions.push(function(){
@@ -257,11 +153,11 @@ test('before cert expiration, fallback to newest cert bundle and stick if node d
 
   // Back to node default
   actions.push(function(){
-    config.instrumental.port = 8000;
+    TestHelper.config.instrumental.port = 8000;
     sendMetric(metricName, realCurrentTime); // bundle fails
   });
   actions.push(function(){
-    config.instrumental.port = 8001;
+    TestHelper.config.instrumental.port = 8001;
     sendMetric(metricName, realCurrentTime); // node default
   });
   actions.push(function(){
@@ -276,7 +172,7 @@ test('before cert expiration, fallback to newest cert bundle and stick if node d
     expectedSum: 6, // failures don't get sent
     found: function(){
       var cert_log_messages =
-        log.filter(function(entry){return entry.match(/\bcert/i)});
+        TestHelper.log.filter(function(entry){return entry.match(/\bcert/i)});
       var expected_messages = [
         "Adding node default ssl cert option",
         'Found valid cert bundle: equifax.2018-08-19',
@@ -309,7 +205,7 @@ test('before cert expiration, fallback to newest cert bundle and stick if node d
     },
     timeout: function(){
       var cert_log_messages =
-        log.filter(function(entry){return entry.match(/Using certs/i)});
+        TestHelper.log.filter(function(entry){return entry.match(/Using certs/i)});
       t.fail(JSON.stringify(cert_log_messages) + "\n\n");
       t.end();
     },
@@ -331,14 +227,14 @@ test('after cert expiration, fallback to newest cert bundle and stick if node de
 
   var metricName = "test.metric"+Math.random();
 
-  config.instrumental.tlsVariationTimeout = 500;
+  TestHelper.config.instrumental.tlsVariationTimeout = 500;
 
   var timeBetweenSends = 1000;
   var actions = [];
 
   // node default certs
   actions.push(function(){
-    config.instrumental.host = "smoke-collector.instrumentalapp.com";
+    TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
     sendMetric(metricName, realCurrentTime); // using node default certs
   });
   actions.push(function(){
@@ -347,11 +243,11 @@ test('after cert expiration, fallback to newest cert bundle and stick if node de
 
   // New cert bundle
   actions.push(function(){
-    config.instrumental.port = 8000;
+    TestHelper.config.instrumental.port = 8000;
     sendMetric(metricName, realCurrentTime); // node default certs fail
   });
   actions.push(function(){
-    config.instrumental.port = 8001;
+    TestHelper.config.instrumental.port = 8001;
     sendMetric(metricName, realCurrentTime); // bundle
   });
   actions.push(function(){
@@ -360,11 +256,11 @@ test('after cert expiration, fallback to newest cert bundle and stick if node de
 
   // Back to node default
   actions.push(function(){
-    config.instrumental.port = 8000;
+    TestHelper.config.instrumental.port = 8000;
     sendMetric(metricName, realCurrentTime); // bundle fails
   });
   actions.push(function(){
-    config.instrumental.port = 8001;
+    TestHelper.config.instrumental.port = 8001;
     sendMetric(metricName, realCurrentTime); // node default
   });
   actions.push(function(){
@@ -379,7 +275,7 @@ test('after cert expiration, fallback to newest cert bundle and stick if node de
     expectedSum: 6, // failures don't get sent
     found: function(){
       var cert_log_messages =
-        log.filter(function(entry){return entry.match(/\bcert/i)});
+        TestHelper.log.filter(function(entry){return entry.match(/\bcert/i)});
       var expected_messages = [
         "Adding node default ssl cert option",
         'Ignoring expired cert bundle: equifax.2018-08-19',
@@ -415,7 +311,7 @@ test('after cert expiration, fallback to newest cert bundle and stick if node de
     },
     timeout: function(){
       var cert_log_messages =
-        log.filter(function(entry){return entry.match(/Using certs/i)});
+        TestHelper.log.filter(function(entry){return entry.match(/Using certs/i)});
       t.fail(JSON.stringify(cert_log_messages) + "\n\n");
       t.end();
     },
@@ -434,7 +330,7 @@ test('old agent works with new elb', function(t) {
   var time = new Date(Date.parse("2019-01-01"));
   timekeeper.travel(time); // Travel to that date.
 
-  config.instrumental.host = "smoke-collector.instrumentalapp.com";
+  TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
   var metricName = "test.metric"+Math.random();
   sendMetric(metricName, oldTime);
 
@@ -464,7 +360,7 @@ test('old agent works with new elb', function(t) {
 //   var time = new Date(Date.parse("2018-01-01"));
 //   timekeeper.travel(time); // Travel to that date.
 //
-//   config.instrumental.host = "collector.instrumentalapp.com";
+//   TestHelper.config.instrumental.host = "collector.instrumentalapp.com";
 //   now = Math.round(new Date().getTime() / 1000);
 //   var metricName = "test.metric"+Math.random();
 //   sendMetric(metricName, realCurrentTime);
@@ -495,20 +391,20 @@ test('old agent works with new elb', function(t) {
 //   var time = new Date(Date.parse("2019-01-01"));
 //   timekeeper.travel(time); // Travel to that date.
 //
-//   config.instrumental.disallowNodeDefaultCerts = true;
-//   config.instrumental.host = "old-elb.instrumentalapp.com";
+//   TestHelper.config.instrumental.disallowNodeDefaultCerts = true;
+//   TestHelper.config.instrumental.host = "old-elb.instrumentalapp.com";
 //   now = Math.round(new Date().getTime() / 1000);
 //   var metricName = "test.metric"+Math.random();
 //   sendMetric(metricName, oldTime);
 //
 //   checkForMetric(metricName, {
 //     found: function(){
-//       t.fail("Future agent shouldn't work with the old ELB, logs: " + log);
+//       t.fail("Future agent shouldn't work with the old ELB, logs: " + TestHelper.log);
 //       t.end();
 //     },
 //     timeout: function(){
 //       var cert_log_messages =
-//         log.filter(function(entry){
+//         TestHelper.log.filter(function(entry){
 //           return entry.match(/\bcert/i) &&
 //             !entry.match(/Error: unable to get/i) &&
 //             !entry.match(/Error: CERT_UNTRUSTED/i) &&
@@ -541,7 +437,7 @@ test('future agent correctly expires cert and works with new elb', function(t) {
   var time = new Date(Date.parse("2019-01-01"));
   timekeeper.travel(time); // Travel to that date.
 
-  config.instrumental.host = "smoke-collector.instrumentalapp.com";
+  TestHelper.config.instrumental.host = "smoke-collector.instrumentalapp.com";
   now = Math.round(new Date().getTime() / 1000);
   var metricName = "test.metric"+Math.random();
   sendMetric(metricName, oldTime);
